requirements.txt
streamlit
pandas
numpy
snowflake-connector-python
scikit-learn
pyod
plotly
openai